added method to create diff image

+++ Steps/Comparer/HammingDistance.cs
@@ -6,11 +6,11 @@ namespace Steps.Comparer
     {
         public static double FindDistance(string hash1, string hash2)
         {
-            int distance = 
+            int distance =
                 hash1.ToCharArray()
                 .Zip(hash2.ToCharArray(), (c1, c2) => new { c1, c2 })
                 .Count(m => m.c1 != m.c2);
-            return distance / hash1.Length;
+            return (double)distance / hash1.Length;
         }
     }
 }
